[frontend] Add a buildCorrelationData() function to Graph.js
This function starts at the current report, and looks at all of the
entites that are "objects" of it. It then filters down to the entities
that contain 2 or more Report edges, for the purpose of building a
correlation graph that diagrams what other reporting is in the instance
that overlaps with the current report. This algorithm is very similar to
the approach used in MISP for its "correlation" graphs, and is probably
the most common visualization I work off of on our team. This is largely
a copy/paste of buildGraphData() from the same file, so there is likely
more tweaking and cleanup work that needs to occur.