[frontend] Add the views for the Knowledge Correlation graph feature
Similar to the new buildCorrelationData() in Graph.js, this also is
largely a copy/paste/modification of the views already rendering the
existing Knowledge Graph feature. Some things not working yet: the
"report graph" toolbar still needs to be specialized for this chart, as
the graph-editing controls no longer apply in this context. The filters
aren't complete yet, and the "3d" graph view also doesn't render
properly yet. That said, this is about 80% there and renders some really
good correlation graphs.