[worker] Expose SSL verification for OpenCTIApiClient
This allows setting `OPENCTI_SSL_VERIFY` environment variable or
opencti:ssl_verify in the YAML configuration filein the OpenCTI Worker
class to enable SSL validation in by the Python client library.
Self-signed certs work if this variable is set to the path to the
signing certificate authority, as currently the client does not respect
`SSL_CERT_FILE` when set on the worker node.