[frontend] In correlations, move filtering into buildCorrelationData()
When applying filters (type, createdBy, markedBy) in the correlation
case, we want to hide any nodes not matching the filters, and also we
want to make sure that any reports that would now be unlinked are also
hidden. This means we need to do another graph discovery process after
every filter change. This moves the filter application code into the
buildCorrelationData() function, stops calling applyFilters() directly
from the component code, and completes a filter application by
discovering a new graphData and setting state with it, plus the filter
modifications.