[frontend] Make the time-range selector adjust the graph
Modify the handlers for the time-range selector such that the time range
will apply only to which reports are rendered on the chart, and
corresponding markers on the timeline will only chart the report nodes.
It may become useful down the line to extend this for other analysis
types, like incident, opinion, note, etc.