[frontend] Disable unused toolbar edit buttons on correlation charts
A number of the buttons that are present on the event graph don't make
sense to include in the correlation charts. For instance: the
relationship editor, and the add/delete features don't belong, as this
chart isn't for editing report-entity relationships. This change will
make these toolbar items only render when handlers are defined for them,
and where there are no handlers defined, they will not even be rendered.
This enables the same toolbar code to be used in both graph types.