Add a "drag to select" feature to Knowledge Graphs
This change adds the react-rectangle-selection widget to the project,
and uses it to implement a rectangular selection-tool. Normal behavior
is to change the current selection to whatever nodes are selected by
the rectangle. If SHIFT is held down, then the nodes within the
selection region will be added to the existing selection. If ALT is
held, then any already-selected nodes within the rectangle will be
removed from the selection.