fix: prevent too_many_clauses error in query builder
By default the elQueryBodyBuilder will use match_phrase for nested query
building. This is fine for data that has fewer than 1024 clauses but
anything larger than that will cause a TooManyClauses error. This is a
problem for objects that have a large number of relationships (>1024).
The fix is to make use of term based queries for keyword fields, which
is what this commit does. For now this will only apply to the core ID
fields as there is no list of all keywordable fields.