Added: pt_br translate for backend and frontend
The translation has been integrated into both the Backend and Frontend of the OpenCTI Platform.
The Backend translation has undergone a comprehensive human review to ensure full technical accuracy and consistency.
The Frontend translation, on the other hand, is currently machine-generated and is in the process of being manually reviewed and refined. Despite being under revision, it is already functional and can be used in its current state.