[frontend] improve widget visualization typing
add explicit type cast to widget types including empty string as the buildConfig function can return an empty widget type. Changing the types everywhere to avoid type casts is possible but a big chunk of work that goes up to graphql definitions so I'll limit my scouting to this for now.